Delete script/tests.js and moving code into run.js 'test' scripts. Resolves Luteceo/aspnet-starter-kit-2.0#4
