+++ run.js
@@ -69,6 +69,43 @@ tasks.set('build', () => {
     }));
 });
 
+
+//
+// Run tests
+// -----------------------------------------------------------------------------
+tasks.set('test', () => {
+  // Do this as the first thing so that any code reading it knows the right env.
+  process.env.BABEL_ENV = 'test';
+  process.env.NODE_ENV = 'test';
+  process.env.PUBLIC_URL = '';
+  return Promise.resolve()
+    //.then(() => run('clean'))
+    //.then(() => run('bundle'))
+    //.then(() => run('copy'))
+    //.then(() => run('appsettings'))
+    .then(() => new Promise((resolve, reject) => {
+      // Makes the script crash on unhandled rejections instead of silently
+      // ignoring them. In the future, promise rejections that are not handled will
+      // terminate the Node.js process with a non-zero exit code.
+      process.on('unhandledRejection', err => {
+        throw err;
+      });
+
+      // Ensure environment variables are read.
+      // require('../config/env');
+
+      const jest = require('jest');
+      const argv = process.argv.slice(2);
+
+      // Watch unless on CI or in coverage mode
+      if (!process.env.CI && argv.indexOf('--coverage') < 0) {
+        argv.push('--watch');
+      }
+
+      jest.run(argv);
+    }));
+});
+
 //
 // Build website and launch it in a browser for testing in watch mode
 // -----------------------------------------------------------------------------
